Pin the libyang version
The apkg build command fails when the commit message for the version-tagged
commit does not start `VERSION`. 

IMHO this is a bug in the libyang apkg config.

To work around it, use an env var (LIBYANG_VERSION) to specify the libyang version 
to be used in order to bypass the buggy script.